Add files via upload
The Recipe Finder is a user-friendly GUI application developed in Python using the tkinter library. This interactive tool allows users to input a list of food items, which the application then employs to query the Edamam API, a robust source for recipe information. The GUI provides a clean interface with an entry field for food items and a button to trigger the recipe search.

The underlying code utilizes the requests library to handle HTTP requests to the Edamam API. The user's input is processed, and relevant parameters are included in the API request, such as the food items and the required API credentials (app ID and key). Upon receiving the API response, the application extracts and displays recipe information.

In terms of skills demonstrated, this project showcases proficiency in GUI development using the tkinter framework, highlighting the ability to create a responsive and intuitive user interface. Additionally, it demonstrates competence in working with external APIs, as evidenced by the use of the requests library to interact with the Edamam API for recipe retrieval.

Overall, the Recipe Finder serves as a practical and engaging tool for individuals seeking culinary inspiration based on their preferred ingredients, embodying the fusion of GUI development, API integration, and Python programming skills.